requirements.txt
# ThreatTrace - Threat Intelligence OSINT
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# Cache
cachetools>=5.3.0

# Rate Limiting
slowapi>=0.1.9

# PDF Export
reportlab>=4.0.0

# Pydantic (incluído com FastAPI, mas explícito)
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0
